internal/helm: move types to internal/types
