+++ internal/helm/controller/reconcile.go
@@ -35,8 +35,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/operator-framework/operator-sdk/internal/helm/internal/diff"
+	"github.com/operator-framework/operator-sdk/internal/helm/internal/types"
 	"github.com/operator-framework/operator-sdk/internal/helm/release"
-	"github.com/operator-framework/operator-sdk/internal/helm/types"
 )
 
 // blank assignment to verify that HelmOperatorReconciler implements reconcile.Reconciler
